Arduino/Teensyduino integration
When the board isn't in "programming mode", the current code
just shows an error and exits. This MR shows a window telling
the user to set their board in "programming mode", and checks
twice a second for the board to appear. When that happens, it
closes the window and proceed to run the desired operation.

This allows to integrate ubaboot.py with arduino/teensyduino
IDE.